Juniper_logprocrssing/requirements.txt
sh
mysql-connector-python
